pyproject.toml
[project]
name = "rect_graph_connector"
version = "0.1.0"
description = "A GUI application for creating and manipulating rectangular graph structures"
authors = [
    {name = "GeminiNinth", email = "gemini.ninth@gmail.com"}
]
dependencies = [
    "PyQt5>=5.15.0",
    "pyyaml>=6.0.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-qt>=4.2.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.1.0",
]

[project.urls]
Repository = "https://github.com/GeminiNinth/rect_graph_connector.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rect_graph_connector"]

[tool.pytest.ini_options]
testpaths = ["src/test"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
